requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
httpx>=0.25.0
boto3>=1.28.62
python-dotenv>=1.0.0
composio-core>=0.3.0
